chore: reduce docker image size
This introduces the use of golang build tags in order to allow pruning builds for specific purposes:

- full
- compile
- observe
- manage

For the docker image, which enables none of those tags, this seems to give us 69Mi -> 19.2Mi
TODO: why is it still big at 19.2Mi?


tags instead
