+++ cmd/subcommands/logs.go
@@ -1,3 +1,5 @@
+//go:build full || observe
+
 package subcommands
 
 import (
